requirements.txt
streamlit
Pillow
vertexai
python-dotenv
google-cloud-aiplatform